Redux
I have been reading up and learning about redux and its place in the react eco system. From what i can gleam it is not used as much in modern day but that a few years ago would have been a pivotal tool for react development. Because of this i felt it prudent i have at least a basic understanding of it in case i encounter it in older code bases down the road.

Thankfully so far it is quite similar to useReducer and Context API so i have gotten started with some extremely basic action creator functions for this simple bank account app, as well as laying down the boiler plate for some of the components.