Legacy Redux
I have been going over a bunch of online materials to learn all i can about redux and its implementation and uses.

I have experimented here with some legacy redux using middleware such as "redux-thunk" to handle asynchronous operations such as the currency conversion API call, which is now handled after the action is dispatched but before reaching the store in the case that it must be first converted.  we also make use of the connect function from redux in the balanceDisplay file to map some state from the store to a prop as more practice with legacy redux.

I understand this is mostly legacy. However older code bases will certainly still make use of it and i felt it important i have a working understanding. I also played around with some of the hooks we now have such as useSelector and useDispatch for a more modern approach and will be getting into react toolkit here starting tomorrow.