Redux Toolkit
Today i have been learning about Redux-Toolkit, and wanted to implement it into this application for the practice.

I have refactored the code to work with redux toolkit. Firstly, i modified the store to use the configureStore function we are given. I then altered my slices to use the createSlice function and refactored my action creators into the new methods of the account slice.

In the case that these methods needed multiple arguments i used the prepare method to prepare the payload into a single value as redux toolkit does not allow for multiple arguments in our action creators, which toolkit now produces for us. Otherwise we can now essentially mutate the data, although it will be converted back to immutable using immer before executing.

For now i have left the asynchronous operation using regular redux and not toolkit as i haven't gotten to that yet but as i understand it redux will still work totally fine with both of these at play.